refactor: split settings helpers
